# src/main/resources/static/docs/openapi3.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "JMT Monster",
    "description" : "Interactive Dining Review Game",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://localhost:8000"
  }, {
    "url" : "https://backend.jmtmonster.com"
  } ],
...
  "paths" : {
...
    "/api/v1/friends" : {
      "get" : {
        "tags" : [ "api" ],
        "summary" : "친구 조회",
        "description" : "친구 목록 전체를 조회합니다.",
        "operationId" : "friend-get-list",
        "parameters" : [ {
          "name" : "from-user-id",
          "in" : "query",
          "description" : "사용자 ID",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "to-user-id",
          "in" : "query",
          "description" : "친구 ID",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "is-hanging-out",
          "in" : "query",
          "description" : "놀러가기 여부",
          "required" : false,
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api-v1-friends-1521885839"
                },
                "examples" : {
                  "friend-get-list" : {
                    "value" : "{\n  \"content\" : [ {\n    \"id\" : \"bdb6c5b2-f211-42ff-9242-8f1840efb484\",\n    \"fromUser\" : {\n      \"id\" : \"135ab1b8-520a-40fb-8423-83d927aefa80\",\n      \"name\" : \"FromUser\",\n      \"email\" : \"test@jmt-monster.com\",\n      \"nickname\" : \"FromUser\",\n      \"imageUrl\" : \"https://profile.example.com/from-user\"\n    },\n    \"toUser\" : {\n      \"id\" : \"3d60987e-b8a4-4ec3-ad9c-f21fa17a64b2\",\n      \"name\" : \"ToUser\",\n      \"email\" : \"test@jmt-monster.com\",\n      \"nickname\" : \"ToUser\",\n      \"imageUrl\" : \"https://profile.example.com/to-user\"\n    },\n    \"isHangingOut\" : false\n  } ],\n  \"pageable\" : {\n    \"sort\" : {\n      \"empty\" : true,\n      \"sorted\" : false,\n      \"unsorted\" : true\n    },\n    \"offset\" : 0,\n    \"pageNumber\" : 0,\n    \"pageSize\" : 10,\n    \"paged\" : true,\n    \"unpaged\" : false\n  },\n  \"last\" : true,\n  \"totalPages\" : 1,\n  \"totalElements\" : 1,\n  \"number\" : 0,\n  \"sort\" : {\n    \"empty\" : true,\n    \"sorted\" : false,\n    \"unsorted\" : true\n  },\n  \"size\" : 10,\n  \"first\" : true,\n  \"numberOfElements\" : 1,\n  \"empty\" : false\n}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/friends/{id}" : {
      "get" : {
        "tags" : [ "api" ],
        "summary" : "친구 조회",
        "description" : "친구 한 건을 조회합니다.",
        "operationId" : "friend-get-one",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "description" : "",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "200",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/api-v1-friends-id-491435824"
                },
                "examples" : {
                  "friend-get-one" : {
                    "value" : "{\n  \"id\" : \"1c990ab7-bd4d-40e0-84c0-80603c43d6d8\",\n  \"fromUser\" : {\n    \"id\" : \"18c3529a-1a8d-4da0-9792-ffb03d238ee9\",\n    \"name\" : \"FromUser\",\n    \"email\" : \"test@jmt-monster.com\",\n    \"nickname\" : \"FromUser\",\n    \"imageUrl\" : \"https://profile.example.com/from-user\"\n  },\n  \"toUser\" : {\n    \"id\" : \"d5997335-2d1d-4b7a-a84f-14a5579b7417\",\n    \"name\" : \"ToUser\",\n    \"email\" : \"test@jmt-monster.com\",\n    \"nickname\" : \"ToUser\",\n    \"imageUrl\" : \"https://profile.example.com/to-user\"\n  },\n  \"isHangingOut\" : false\n}"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "api-v1-friends-id-491435824" : {
        "type" : "object",
        "properties" : {
          "toUser" : {
            "type" : "object",
            "properties" : {
              "imageUrl" : {
                "type" : "string",
                "description" : "친구 프로필 사진 주소"
              },
              "nickname" : {
                "type" : "string",
                "description" : "친구 닉네임"
              },
              "name" : {
                "type" : "string",
                "description" : "친구 이름"
              },
              "id" : {
                "type" : "string",
                "description" : "친구 ID"
              },
              "email" : {
                "type" : "string",
                "description" : "친구 이메일"
              }
            }
          },
          "isHangingOut" : {
            "type" : "boolean",
            "description" : "놀러가기 여부"
          },
          "fromUser" : {
            "type" : "object",
            "properties" : {
              "imageUrl" : {
                "type" : "string",
                "description" : "사용자 프로필 사진 주소"
              },
              "nickname" : {
                "type" : "string",
                "description" : "사용자 닉네임"
              },
              "name" : {
                "type" : "string",
                "description" : "사용자 이름"
              },
              "id" : {
                "type" : "string",
                "description" : "사용자 ID"
              },
              "email" : {
                "type" : "string",
                "description" : "사용자 이메일"
              }
            }
          },
          "id" : {
            "type" : "string",
            "description" : "ID"
          }
        }
      },
...
      "api-v1-friends-1521885839" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "type" : "object",
              "properties" : {
                "toUser" : {
                  "type" : "object",
                  "properties" : {
                    "imageUrl" : {
                      "type" : "string",
                      "description" : "친구 프로필 사진 주소"
                    },
                    "nickname" : {
                      "type" : "string",
                      "description" : "친구 닉네임"
                    },
                    "name" : {
                      "type" : "string",
                      "description" : "친구 이름"
                    },
                    "id" : {
                      "type" : "string",
                      "description" : "친구 ID"
                    },
                    "email" : {
                      "type" : "string",
                      "description" : "친구 이메일"
                    }
                  }
                },
                "isHangingOut" : {
                  "type" : "boolean",
                  "description" : "놀러가기 여부"
                },
                "fromUser" : {
                  "type" : "object",
                  "properties" : {
                    "imageUrl" : {
                      "type" : "string",
                      "description" : "사용자 프로필 사진 주소"
                    },
                    "nickname" : {
                      "type" : "string",
                      "description" : "사용자 닉네임"
                    },
                    "name" : {
                      "type" : "string",
                      "description" : "사용자 이름"
                    },
                    "id" : {
                      "type" : "string",
                      "description" : "사용자 ID"
                    },
                    "email" : {
                      "type" : "string",
                      "description" : "사용자 이메일"
                    }
                  }
                },
                "id" : {
                  "type" : "string",
                  "description" : "ID"
                }
              }
            }
          }
        }
      },
...
    }
  }
}